app/requirements.txt
numpy
pandas
matplotlib
pillow
scikit-learn
streamlit
tensorflow